refactor(entity): Refactor to use BaseEvent base class
