+++ src/Entity/SolutionVideoEvent.php
@@ -4,15 +4,12 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
-use App\Entity\Trait\WithUlid;
 use App\Repository\SolutionVideoEventRepository;
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity(repositoryClass: SolutionVideoEventRepository::class)]
-class SolutionVideoEvent
+class SolutionVideoEvent extends BaseEvent
 {
-    use WithUlid;
-
     #[ORM\ManyToOne(inversedBy: 'solutionVideoEvents')]
     #[ORM\JoinColumn(nullable: false)]
     private ?User $opener = null;
